# swagger.yaml
openapi: 3.0.0
paths:
...
  /promocodes:
    post:
      operationId: PromocodesController_create
      summary: Добавить новый промокод
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromocodeDto'
      responses:
        '201':
          description: Промокод успешно добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promocode'
        '403':
          description: Отказ в доступе
        '409':
          description: Такой промокод уже существует
      tags: &ref_12
        - promocodes
      security: &ref_13
        - bearer: []
    get:
      operationId: PromocodesController_findAll
      summary: Получить все промокоды
      parameters: []
      responses:
        '200':
          description: Промокоды успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Promocode'
        '403':
          description: Отказ в доступе
      tags: *ref_12
      security: *ref_13
  /promocodes/promocode:
    get:
      operationId: PromocodesController_findOneByCode
      summary: Получить промокод по названию
      parameters:
        - name: code
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Промокод успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Promocode'
        '403':
          description: Отказ в доступе
      tags: *ref_12
      security: *ref_13
    patch:
      operationId: PromocodesController_updateByCode
      summary: Использовать промокод 1 раз
      parameters:
        - name: code
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Промокод успешно обновлен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Promocode'
        '403':
          description: Отказ в доступе
      tags: *ref_12
      security: *ref_13
  /promocodes/{id}:
    get:
      operationId: PromocodesController_findOne
      summary: Получить промокод по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор промокода
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Промокод успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promocode'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_12
      security: *ref_13
    patch:
      operationId: PromocodesController_update
      summary: Изменить данные промокода по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор промокода
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePromocodeDto'
      responses:
        '200':
          description: Данные промокода успешно изменены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promocode'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_12
      security: *ref_13
    delete:
      operationId: PromocodesController_remove
      summary: Удалить промокод по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор промокода
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Промокод успешно удален
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promocode'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_12
      security: *ref_13
...
  /subscriptions:
    get:
      operationId: SubscriptionsController_subscriptionAndPayments
      summary: Данные страницы "Подписка и платежи"
      parameters: []
      responses:
        '200':
          description: Данные подписок и платежей успешно получены
          content:
            application/json:
              schema:
                type: object
                properties:
                  tariff:
                    type: string
                    example: Бизнес
                  status:
                    type: boolean
                    example: true
                  cardMask:
                    type: string
                    example: 4500 *** 1119
                  debitDate:
                    type: date
                    example: '2023-09-12'
                  balance:
                    type: number
                    example: 1234
                  payments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: &ref_18
        - subscriptions
      security: &ref_19
        - bearer: []
  /subscriptions/activate:
    post:
      operationId: SubscriptionsController_activateSubscription
      summary: Активировать(отменить) подписку
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: boolean
                  example: true
      responses:
        '201':
          description: Подписка активирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_18
      security: *ref_19
  /subscriptions/{id}:
    post:
      operationId: SubscriptionsController_createSubscription
      summary: Оформить подписку
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор тарифа
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cardMask:
                  type: string
                  example: 4500 *** 1119
                debitDate:
                  type: date
                  example: '2023-09-12'
      responses:
        '201':
          description: Подписка оформлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_18
      security: *ref_19
...
info:
  title: API BotKits
  description: Ручки для команды frontend
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreatePromocodeDto:
      type: object
      properties:
        code:
          type: string
          example: PROMO50
        actionPeriod:
          format: date-time
          type: string
          example: '2023-09-10T16:07:34.285Z'
        activationCount:
          type: number
          minimum: 0
          example: 2
        maxActivationCount:
          type: number
          minimum: 1
          example: 10
        amount:
          type: number
          minimum: 1
          example: 1500
      required:
        - code
        - actionPeriod
        - activationCount
        - maxActivationCount
        - amount
    Promocode:
      type: object
      properties:
        code:
          type: string
          example: PROMO50
        actionPeriod:
          format: date-time
          type: string
          example: '2023-09-10T16:07:34.285Z'
        activationCount:
          type: number
          example: 2
        maxActivationCount:
          type: number
          example: 10
        amount:
          type: number
          example: 1500
      required:
        - code
        - actionPeriod
        - activationCount
        - maxActivationCount
        - amount
    UpdatePromocodeDto:
      type: object
      properties: {}
...
    Subscription:
      type: object
      properties:
        status:
          type: boolean
          example: true
          default: false
        cardMask:
          type: string
          example: 4500 *** 1119
        debitDate:
          format: date-time
          type: string
          example: 2023-10-08T16:43:56.196Z
      required:
        - status
        - cardMask
        - debitDate
    Payment:
      type: object
      properties:
        date:
          format: date-time
          type: string
          example: 2023-10-08T16:43:56.198Z
        amount:
          type: number
          example: 1000
        successful:
          type: boolean
          example: true
        operation:
          type: string
          enum:
            - Поступление
            - Списание
          example: Поступление
        note:
          type: string
          example: Пополнение счета
          default: '-'
      required:
        - date
        - amount
        - successful
        - operation
        - note
...